appflask/requirements.txt
dataclasses==0.7
marshmallow-dataclass==8.1.0
openpyxl==3.0.5
xlrd==1.2.0
docx==0.2.4
pdf2image==1.13.1
PyJWT==1.7.1
deepdiff==5.0.1
pypng==0.0.20
watchdog==0.10.3
pyqrcode==1.2.1
pyzbar==0.1.8
weasyprint==51
usaddress-scourgify==0.2.4
Pillow==7.1.2
geoalchemy2==0.7.0
geopandas==0.6.2
pandas==0.25.3
pandas_access==0.0.1
Babel==2.7.0
blinker==1.4
Click==7.0
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Babel==0.12.2
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
Flask-Migrate==2.5.2
flask-marshmallow==0.10.1
flask-restful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
passlib==1.7.1
pdftotext==2.1.2
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.36.0
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
fuzzywuzzy==0.17.0
python-Levenshtein==0.12.0
rq==1.2.0
tqdm==4.41.1
pyproj==1.9.6
simplejson==3.17.0
Flask-HTTPAuth==3.3.0
Flask-Cors==3.0.8
Flask-Reuploaded==0.3.2
Flask-WeasyPrint==0.6
paginate==0.5.6
geoalchemy2==0.7.0
wget==3.2
orjson==3.2.0
gunicorn==20.0.4
pdfkit==0.6.1
locust==1.1.1
pypdftk==0.4
reportlab==3.5.47
PyPDF2==1.26.0
selenium==3.141.0
lxml==4.5.2
chromedriver-binary==85.0.4183.87.0
comtypes==1.1.7
docxtpl==0.10.0
python-docx==0.8.10
requests==2.22.0
webdriver-manager==3.2.2
shapely-geojson==0.0.1
numpy==1.18.5